god dammit

+++ ml/layout_v1/dataset.py
@@ -413,7 +413,7 @@ class LayoutDataset(Dataset):
         )
         np.save(
             os.path.join(target_file_path, self.CONFIG_FEATURES_FILE),
-            graph_data.config_features.astype(np.float16),
+            graph_data.config_features.astype(np.float32),
         )
         np.save(
             os.path.join(target_file_path, self.CONFIG_RUNTIME_FILE),
